refactor: Remove unused code
