+++ event-handlers/src/main/java/no/sikt/nva/nvi/events/evaluator/calculator/PointCalculationConstants.java
@@ -1,6 +1,10 @@
 package no.sikt.nva.nvi.events.evaluator.calculator;
 
-import static no.sikt.nva.nvi.events.evaluator.model.InstanceType.*;
+import static no.sikt.nva.nvi.common.service.model.InstanceType.ACADEMIC_ARTICLE;
+import static no.sikt.nva.nvi.common.service.model.InstanceType.ACADEMIC_CHAPTER;
+import static no.sikt.nva.nvi.common.service.model.InstanceType.ACADEMIC_COMMENTARY;
+import static no.sikt.nva.nvi.common.service.model.InstanceType.ACADEMIC_LITERATURE_REVIEW;
+import static no.sikt.nva.nvi.common.service.model.InstanceType.ACADEMIC_MONOGRAPH;
 import static no.sikt.nva.nvi.events.evaluator.model.Level.LEVEL_ONE;
 import static no.sikt.nva.nvi.events.evaluator.model.Level.LEVEL_TWO;
 import static no.sikt.nva.nvi.events.evaluator.model.PublicationChannel.JOURNAL;
@@ -10,7 +14,7 @@ import java.math.BigDecimal;
 import java.math.MathContext;
 import java.math.RoundingMode;
 import java.util.Map;
-import no.sikt.nva.nvi.events.evaluator.model.InstanceType;
+import no.sikt.nva.nvi.common.service.model.InstanceType;
 import no.sikt.nva.nvi.events.evaluator.model.Level;
 import no.sikt.nva.nvi.events.evaluator.model.PublicationChannel;
