+++ event-handlers/src/main/java/no/sikt/nva/nvi/events/evaluator/model/PointCalculation.java
@@ -3,7 +3,7 @@ package no.sikt.nva.nvi.events.evaluator.model;
 import java.math.BigDecimal;
 import java.net.URI;
 import java.util.List;
-import java.util.Map;
+import no.sikt.nva.nvi.common.service.model.InstanceType;
 
 public record PointCalculation(InstanceType instanceType,
                                PublicationChannel channelType,
