+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/service/model/InstanceType.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.events.evaluator.model;
+package no.sikt.nva.nvi.common.service.model;
 
 import java.util.Arrays;
